Fix: 커스텀훅 추가, 빌드 에러 해결
